ttree - Split option to preserve ownership and permission
Unless you're running as root or the original user of a file, it will be
rather hard to preserve the ownership. Currently using another users
directory as source to create your own target directory comes with a
flood of warnings, which are harmless, but annoying. To be able to get
rid of these warnings we split the --preserve option into
--preserve_mode and --preserve_owner, which allows us to activate
preservation of permission and ownership separately. We adjust the
--preserve option to set both --preserve_mode and --preserve_owner for
backward compatibility reasons.